Rework documentation of the file and repository naming scheme
This quite a large rework.

There is now three documents:

1. Naming_Codes.md
This document lists the codes that are used for the filename.

2. Filename_Structure.md
This document details the encoding of the codes into the filename. With many examples.

3. Repository_Structures.md
This document details the folder structure of the Art Repository.
(This document needs to be expanded upon in future commits.)